+++ src/routes/graphql/loaders/postsLoader.ts
